+++ packages/quiz_repository/lib/src/models/achievement_local_model.dart
@@ -6,7 +6,6 @@ class AchievementLocalModel {
   }
 
   final String id;
-  //TODO : int / double
-  final double star;
+  final int star;
   final int hightscore;
 }
